fix compiler warning
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@235 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/xmlrpc-c/src/xmlrpc_server_abyss.c
@@ -770,7 +770,7 @@ normalLevelAbyssRun(xmlrpc_env *                      const envP ATTR_UNUSED,
                     const xmlrpc_server_abyss_parms * const parmsP,
                     unsigned int                      const parm_size) {
     
-    uint16 portNumber;
+    unsigned int portNumber;
     
     DateInit();
     MIMETypeInit();
